[del/#217] SignOutNavigation.kt 파일 제거
- mypageGroup에 navController 매개변수 추가
