+++ feature/main/src/main/java/com/napzak/market/main/MainScreen.kt
@@ -41,9 +41,6 @@ import com.napzak.market.main.component.MainBottomBar
 import com.napzak.market.main.component.MainRegisterDialog
 import com.napzak.market.main.component.MainSnackBarHost
 import com.napzak.market.mypage.navigation.mypageGraph
-import com.napzak.market.mypage.navigation.navigateToSettings
-import com.napzak.market.mypage.signout.navigation.navigateToSignOut
-import com.napzak.market.mypage.signout.navigation.signOutGraph
 import com.napzak.market.onboarding.navigation.Terms
 import com.napzak.market.onboarding.navigation.onboardingGraph
 import com.napzak.market.registration.navigation.navigateToGenreSearch
@@ -296,22 +293,15 @@ private fun MainNavHost(
         )
 
         mypageGraph(
-            restartApplication = restartApplication,
+            navController = navigator.navController,
             navigateToUp = navigator::navigateUp,
             navigateToMyMarket = { navigator.navController.navigateToStore(it) },
             navigateToSales = { /* TODO: 판매내역 화면으로 이동 */ },
             navigateToPurchase = { /* TODO: 구매내역 화면으로 이동 */ },
             navigateToRecent = { /* TODO: 최근 본 상품 화면으로 이동 */ },
             navigateToFavorite = { /* TODO: 찜 화면으로 이동 */ },
-            navigateToSettings = navigator.navController::navigateToSettings,
-            navigateToSignOut = navigator.navController::navigateToSignOut,
+            restartApplication = restartApplication,
             modifier = modifier,
-        )
-
-        signOutGraph(
-            navController = navigator.navController,
-            onNavigateUp = navigator::navigateUp,
-            restartApp = restartApplication,
         )
     }
 }
